+++ src/server_manager/web_app/survey.ts
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-import {sleep} from '../infrastructure/async';
+import {sleep} from '../infrastructure/sleep';
 import {Surveys} from '../model/survey';
 
 export const DEFAULT_PROMPT_IMPRESSION_DELAY_MS = 3000;
